Extracted the part that holds the wakelock to an actual service.
The service is started when the caffeine tile is added, and stopped when it is removed. Whenever the tile needs to know or toggle the state of the wakelock, it communicates with the WakelockService through an IBinder.

Changed the source/target language to Java 8, and enabled the new Jack toolchain in order to have it all work. End result: much smaller APK!